+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Changelog
 ------------------
 * Fixed making dataset as aminoacid seqs for MEGA format.
 * Fixed making dataset as degenerated seqs for MEGA format.
+* Raise ValueError when asked to degenerate seqs that will go to partitioning
+  based on codon positions.
 
 0.3.7 (2015-10-30)
 ------------------